Fix build

+++ utils/funcexp/functor_bool.h
@@ -353,7 +353,7 @@ class Func_Compare : public Func_Bool
     return resultType;
   }
 
-  int getIntVal(rowgroup::Row& row, FunctionParm& fp, bool& isNull,
+  int64_t getIntVal(rowgroup::Row& row, FunctionParm& fp, bool& isNull,
                   execplan::CalpontSystemCatalog::ColType& op_ct)
   {
     return getBoolVal(row, fp, isNull, op_ct);
@@ -487,7 +487,7 @@ class Func_Logic_Op : public Func_Bool
     return resultType;
   }
 
-  int getIntVal(rowgroup::Row& row, FunctionParm& fp, bool& isNull,
+  int64_t getIntVal(rowgroup::Row& row, FunctionParm& fp, bool& isNull,
                   execplan::CalpontSystemCatalog::ColType& op_ct)
   {
     return getBoolVal(row, fp, isNull, op_ct);
